// react-burger/src/types/ingredient.ts
export interface Ingredient {
  _id: string
  name: string
  type: 'bun' | 'sauce' | 'main'
  proteins: number
  fat: number
  carbohydrates: number
  calories: number
  price: number
  image: string
  image_mobile: string
  image_large: string
  __v: number
}

export interface ConstructorIngredient extends Ingredient {
  uuid: string
}